Rename the package from ...newtask to ...task.
