+++ android-app/app/src/main/java/io/spine/examples/todolist/task/ExistingLabelsAdapter.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.todolist.newtask;
+package io.spine.examples.todolist.task;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
